% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanteamghana.R
\docType{data}
\name{cleanteamghana}
\alias{cleanteamghana}
\title{Ghana Clean Team Survey}
\format{
A tibble with 584 rows and 4 variables:
\describe{
\item{age}{ Age of repondent}
\item{area}{ Origin area in Ghana from the respondent}
\item{toilet}{ Question: Are you able to observe a toilet?}
\item{money_toilet}{ Are you able to afford a toilet?}
}
}
\usage{
cleanteamghana
}
\description{
ADD DESCRIPTION
TBD
}
\keyword{datasets}
